backend/requirements.txt
Django==4.2
djangorestframework==3.15.2
django-colorfield
django-cors-headers==4.6.0
django-extensions==3.2.3
django-filter==24.3
django-rest-knox==5.0.2
Pillow==11.0.0
requests==2.32.3

# asynchronous tasks
celery~=5.4.0
redis~=5.2.0

# social auth
social-auth-app-django==5.4.2

# payments/subscriptions
stripe==11.2.0

# error tracking
sentry-sdk[django]~=2.18.0

# testing
Faker==30.8.2

# dev dependencies
flake8
isort
black
bandit
coverage

# production
psycopg2-binary==2.9.10
uwsgi
